fix int price -> long price

+++ src/main/java/ptithcm/dao/promotion/PromotionDao.java
@@ -7,7 +7,7 @@ import ptithcm.model.promotion.Promotion;
 import ptithcm.model.promotion.PromotionCategory;
 
 public interface PromotionDao {
-	public long getPriceDiscount(int productItemId, Integer oldPrice);
+	public long getPriceDiscount(int productItemId, long oldPrice);
 	public List<Promotion> getListPromotion();
 	public Promotion getPromotionById(int promotionId);
 	public Integer getProductCategoryId(int promotionId);
