+++ src/main/java/ptithcm/service/admin/PromotionService.java
@@ -17,7 +17,7 @@ public class PromotionService {
 	@Autowired
 	PromotionDao promotionDao;
 	
-	public long getPriceDiscount(int productItemId, Integer oldPrice) {
+	public long getPriceDiscount(int productItemId, long oldPrice) {
 		long price = promotionDao.getPriceDiscount(productItemId, oldPrice);
 		if (price >= 0) {
 			System.out.println("percent: " + price);
